Split of responsibilities between Inventory, InventoryController, and InventoryPanelController
