+++ Assets/Scenes/Main.unity
@@ -15460,6 +15460,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   playerInventory: {fileID: 1559548625}
+  inventoryController: {fileID: 1417121306}
   slotPrefab: {fileID: 1296966723602450390, guid: 4ed57193ddf16034bb67688966b3108f, type: 3}
 --- !u!1 &1163910803 stripped
 GameObject:
@@ -18761,6 +18762,7 @@ GameObject:
   - component: {fileID: 1417121304}
   - component: {fileID: 1417121303}
   - component: {fileID: 1417121302}
+  - component: {fileID: 1417121306}
   m_Layer: 5
   m_Name: Inventory UI
   m_TagString: Untagged
@@ -18852,6 +18854,19 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!114 &1417121306
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1417121301}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 649a8d4b7010a404e814f4e66205559d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: Assembly-CSharp::InventoryController
+  playerInventory: {fileID: 1559548625}
 --- !u!4 &1420181651 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4461063578722904, guid: 204618b63d9313044aeaf24f717cf848, type: 3}
@@ -20668,7 +20683,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   inventorySize: 30
-  itemCursorFollowerController: {fileID: 1451227154}
 --- !u!114 &1559548626
 MonoBehaviour:
   m_ObjectHideFlags: 0
