+++ Assets/Scripts/Inventory/Inventory.cs
@@ -7,8 +7,6 @@ public class Inventory : MonoBehaviour
 {
     [SerializeField] private int inventorySize = 30;
 
-    private Dictionary<InventorySlotUIController, int> slotUIControllerToIndexMap;
-
     public class InventoryEntry
     {
         public Item item { get; }
@@ -73,91 +71,14 @@ public class Inventory : MonoBehaviour
     private void Awake()
     {
         inventoryEntries = new InventoryEntry[inventorySize];
-        slotUIControllerToIndexMap = new Dictionary<InventorySlotUIController, int>();
     }
 
     private void Update()
     {
     }
-
-    private void OnEnable()
+    public InventoryOperationResult InteractWithSlot(int index)
     {
-        EventManager.InventorySlotClickedEvent += HandleUIInventorySlotClicked;
-    }
-
-    private void OnDisable()
-    {
-        EventManager.InventorySlotClickedEvent -= HandleUIInventorySlotClicked;
-    }
-
-    private void RefreshSlots(InventoryOperationResult inventoryOperationResult)
-    {
-        if (inventoryOperationResult.OperationResultType == InventoryOperationResult.ResultType.NoOperation || (!inventoryOperationResult.CursorSlotChanged && inventoryOperationResult.ChangedSlotIndices.Count == 0))
-            return;
-
-        if (!inventoryOperationResult.CursorSlotChanged)
-            RefreshSlots(inventoryOperationResult.ChangedSlotIndices.ToArray());
-        else
-            RefreshSlots(inventoryOperationResult.ChangedSlotIndices.Prepend(CursorSlotIndex).ToArray());
-    }
-
-    private void RefreshSlots(params int[] indices)
-    {
-        EventManager.TriggerInventoryUpdateEvent(indices);
-    }
-    private void HandleUIInventorySlotClicked(InventorySlotUIController controller)
-    {
-        if (!slotUIControllerToIndexMap.TryGetValue(controller, out int index))
-        {
-            Debug.LogWarning("Clicked slot not registered with Inventory!");
-            return;
-        }
-
-        InventoryOperationResult inventoryOperationResult;
-
-        // Special keypresses: ctrl-click to pick 1 item from the stack, shift-click to pick up a specified number from the stack
-        if (controller.SlotClickType == InventorySlotUIController.ClickType.Ctrl && !ItemInCursorSlot)
-            inventoryOperationResult = HandleUIInventorySlotClickedForSinglePull(index);
-        else if (controller.SlotClickType == InventorySlotUIController.ClickType.Shift && !ItemInCursorSlot)
-        {
-            HandleUIInventorySlotClickedForStackSelection(controller, index);
-            return;
-        }
-        else
-            inventoryOperationResult = HandleInventorySlotInteraction(index);
-
-        RefreshSlots(inventoryOperationResult);
-    }
-
-    private InventoryOperationResult HandleUIInventorySlotClickedForSinglePull(int index)
-    {
-        return TakeFromSlotIntoCursor(index, 1);
-    }
-
-    private void HandleUIInventorySlotClickedForStackSelection(InventorySlotUIController controller, int index)
-    {
-        Action<int> acceptButtonAction = (amountTaken) => OnStackSizeSelectorAccepted(index, amountTaken);
-
-        // Get location for stack size selector panel to open, then open it
-        Vector2 stackSizeSelectorPanelPosition = CalculateStackSelectionPanelPosition(controller);
-        UIManager.Instance.ShowStackSizeSelectorPanel(GetSlotDisplayInformation(index), stackSizeSelectorPanelPosition, acceptButtonAction);
-    }
-
-    private Vector2 CalculateStackSelectionPanelPosition(InventorySlotUIController controller)
-    {
-        RectTransform inventorySlotRectTransform = controller.GetComponent<RectTransform>();
-        Vector2 slotTransformSize = inventorySlotRectTransform.rect.size;  // (width, height) pair for the slot's size
-        Vector2 slotTransformPivotOffset = inventorySlotRectTransform.pivot;  // Represents the pivot offset within the slot transform as a pair (x-offset, y-offset). The values range from 0 to 1, where 0 is the left (x) or bottom (y).
-        Vector2 bottomRightOffset = new Vector2((1.0f - slotTransformPivotOffset.x) * slotTransformSize.x, (0.0f - slotTransformPivotOffset.y) * slotTransformSize.y); // Calculate the offset of the bottom-right corner from the pivot location
-        RectTransform inventoryPanelRectTransform = inventorySlotRectTransform.GetComponentInParent<InventoryPanelController>().GetComponent<RectTransform>();
-        Vector2 bottomRightCorner = inventoryPanelRectTransform.anchoredPosition + inventorySlotRectTransform.anchoredPosition + bottomRightOffset;
-        return bottomRightCorner;
-    }
-
-    private void OnStackSizeSelectorAccepted(int index, int amountTaken)
-    {
-        InventoryOperationResult inventoryOperationResult = TakeFromSlotIntoCursor(index, amountTaken);
-        RefreshSlots(inventoryOperationResult);
+        return HandleInventorySlotInteraction(index);
     }
 
     private InventoryOperationResult HandleInventorySlotInteraction(int index)
@@ -207,7 +128,7 @@ public class Inventory : MonoBehaviour
         }
     }
 
-    private InventoryOperationResult TakeFromSlotIntoCursor(int index, int amount)
+    public InventoryOperationResult TakeFromSlotIntoCursor(int index, int amount)
     {
         if (index < 0 || index >= inventoryEntries.Length)
             throw new ArgumentOutOfRangeException(nameof(index), $"Inventory index must be between 0 and {inventoryEntries.Length - 1}; index given was {index}");
@@ -392,25 +313,6 @@ public class Inventory : MonoBehaviour
         return InventoryOperationResult.ItemFullyAdded(indicesToUpdate.ToArray());
     }
 
-    public InventoryOperationResult HandleAddItemToInventory(Item item, int amountToAdd, int? slotIndexChoice = null, bool allowOverflowOutsideChosenSlot = true)
-    {
-        if (item == null)
-        {
-            Debug.LogWarning("Tried to add a null item to the inventory.");
-            return InventoryOperationResult.NoOperation();
-        }
-        InventoryOperationResult inventoryOperationResult = AddItem(item, amountToAdd, slotIndexChoice, allowOverflowOutsideChosenSlot);
-        if (inventoryOperationResult.OperationResultType == InventoryOperationResult.ResultType.ItemPartiallyAdded || inventoryOperationResult.OperationResultType == InventoryOperationResult.ResultType.NoSpace)
-            Debug.LogWarning($"Out of {amountToAdd} items to add, {inventoryOperationResult.LeftoverItemCount} were lost due to not having space in the inventory. (Overflow to other slots was {(allowOverflowOutsideChosenSlot ? "" : "not ")}enabled.)");
-        RefreshSlots(inventoryOperationResult);
-        return inventoryOperationResult;
-    }
-
-    public void RegisterUIInventorySlot(InventorySlotUIController slotUIController, int index)
-    {
-        slotUIControllerToIndexMap[slotUIController] = index;
-    }
-
     public InventoryOperationResult PutItemInCursorSlot(Item item, int amount)
     {
         if (item == null)
@@ -426,23 +328,6 @@ public class Inventory : MonoBehaviour
         cursorInventoryEntry = new InventoryEntry(item, amount);
         InventoryOperationResult result = InventoryOperationResult.PickupToCursor(null);
         return result;
-    }
-
-    public InventoryOperationResult HandlePutItemInCursorSlot(Item item, int amount)
-    {
-        if (item == null)
-        {
-            Debug.LogWarning("Tried to add a null item to the cursor.");
-            return InventoryOperationResult.NoOperation();
-        }
-        InventoryOperationResult inventoryOperationResult = PutItemInCursorSlot(item, amount);
-        if (inventoryOperationResult.OperationResultType == InventoryOperationResult.ResultType.NoSpace)
-            Debug.LogWarning("Tried to put something in the cursor inventory slot while it already had something in it.");
-        else if (inventoryOperationResult.OperationResultType == InventoryOperationResult.ResultType.NoOperation || !inventoryOperationResult.CursorSlotChanged)
-            Debug.LogWarning("No change when attempting to add an item to the cursor slot.");
-
-        RefreshSlots(inventoryOperationResult);
-        return inventoryOperationResult;
     }
 
     public InventorySlotDisplayInformation GetSlotDisplayInformation(int index)
