+++ Assets/Scripts/Inventory/InventoryController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 649a8d4b7010a404e814f4e66205559d
